Move tests out of TestChineseUtils
