+++ src/jyut-dict/CMakeLists.txt
@@ -176,7 +176,9 @@ add_subdirectory(logic/search/test/TestSqlSearch)
 add_subdirectory(logic/sentence/test/TestSentenceSet)
 add_subdirectory(logic/sentence/test/TestSourceSentence)
 add_subdirectory(logic/settings/test/TestSettingsUtils)
+add_subdirectory(logic/utils/test/TestCantoneseUtils)
 add_subdirectory(logic/utils/test/TestChineseUtils)
+add_subdirectory(logic/utils/test/TestMandarinUtils)
 add_subdirectory(logic/utils/test/TestScriptDetector)
 
 set_target_properties(CantoneseDictionary PROPERTIES
