+++ src/jyut-dict/logic/utils/test/TestChineseUtils/tst_chineseutils.cpp
@@ -1,7 +1,6 @@
 #include <QtTest>
 
 #include "logic/settings/settings.h"
-#include "logic/utils/cantoneseutils.h"
 #include "logic/utils/chineseutils.h"
 #include "logic/utils/mandarinutils.h"
 #include "logic/utils/utils.h"
@@ -24,119 +23,10 @@ private slots:
     void compareStringsMultibyteGraphemesWithAlpha();
     void compareStringsCompatibilityVariantNormalization();
 
-    void jyutpingToYaleSimple();
-    void jyutpingToYaleRejectNoTone();
-    void jyutpingToYaleRejectSingleLetter();
-    void jyutpingToYaleRejectSpecialCharacter();
-    void jyutpingToYaleNoSpaces();
-    void jyutpingToYaleSpacesToSegment();
-    void jyutpingToYaleSpecialFinal();
-    void jyutpingToYaleLightTone();
-    void jyutpingToYaleSpecialSyllable();
-    void jyutpingToYaleTones();
-    void jyutpingToYaleNoTone();
-
-    void jyutpingToIPASimple();
-    void jyutpingToIPARejectNoTone();
-    void jyutpingToIPARejectSingleLetter();
-    void jyutpingToIPARejectSpecialCharacter();
-    void jyutpingToIPANoSpaces();
-    void jyutpingToIPASpacesToSegment();
-    void jyutpingToIPAPreprocessInitial();
-    void jyutpingToIPASpecialSyllable();
-    void jyutpingToIPACheckedTone();
-    void jyutpingToIPASpecialFinal();
-    void jyutpingToIPATones();
-    void jyutpingToIPANoTone();
-
-    void prettyPinyinSimple();
-    void prettyPinyinRejectNoTone();
-    void prettyPinyinRejectSingleLetter();
-    void prettyPinyinRejectSpecialCharacter();
-    void prettyPinyinSecondaryVowel();
-    void prettyPinyinUmlaut();
-    void prettyPinyinTones();
-    void prettyPinyinNoTone();
-
-    void numberedPinyinSimple();
-
-    void pinyinWithVSimple();
-
-    void pinuyinToZhuyinSimple();
-    void pinuyinToZhuyinRejectNoTone();
-    void pinuyinToZhuyinRejectSingleLetter();
-    void pinuyinToZhuyinRejectSpecialCharacter();
-    void pinuyinToZhuyinNoSpaces();
-    void pinuyinToZhuyinSpacesToSegment();
-    void pinuyinToZhuyinSpecialInitial();
-    void pinuyinToZhuyinSpecialFinals();
-    void pinuyinToZhuyinErhua();
-    void pinuyinToZhuyinMalformed();
-
-    void pinyinToIPASimple();
-    void pinyinToIPARejectNoTone();
-    void pinyinToIPARejectSingleLetter();
-    void pinyinToIPARejectSpecialCharacter();
-    void pinyinToIPANoSpaces();
-    void pinyinToIPASpacesToSegment();
-    void pinyinToIPASpecialCaseNg();
-    void pinyinToIPASpecialCaseRi();
-    void pinyinToIPASyllableWithV();
-    void pinyinToIPAVoicelessInitial();
-    void pinyinToIPAToneThree();
-    void pinyinToIPAToneFour();
-    void pinyinToIPAOtherTone();
-    void pinyinToIPAErhua();
-
     void constructRomanisationQuerySingleSyllable();
     void constructRomanisationQueryMultiSyllable();
     void constructRomanisationQueryGlobCharacters();
     void constructRomanisationQueryOnlyGlobCharacters();
-
-    void segmentJyutpingSimple();
-    void segmentJyutpingNoDigits();
-    void segmentJyutpingNoSpaces();
-    void segmentJyutpingNoDigitsNoSpaces();
-    void segmentJyutpingNoDigitsApostrophe();
-    void segmentJyutpingDigitsApostrophe();
-    void segmentJyutpingRemoveSpecialCharacters();
-    void segmentJyutpingKeepGlobCharacters();
-    void segmentJyutpingKeepGlobCharactersNoWhitespace();
-    void segmentJyutpingKeepMultipleGlobCharacters();
-    void segmentJyutpingKeepMultipleGlobCharactersWhitespace();
-    void segmentJyutpingKeepMultipleGlobCharactersWhitespaceSurround();
-    void segmentJyutpingGlobCharactersTrimWhitespace();
-    void segmentJyutpingKeepSpecialCharacters();
-    void segmentJyutpingRemoveWhitespace();
-    void segmentJyutpingLower();
-    void segmentJyutpingLowerWithDigits();
-    void segmentJyutpingMultipleFinalsVowelsOnly();
-    void segmentJyutpingMultipleFinalsConsonantsOnly();
-    void segmentJyutpingMultipleFinals();
-    void segmentJyutpingInvalidTone();
-    void segmentJyutpingGarbage();
-
-    void segmentPinyinSimple();
-    void segmentPinyinNoDigits();
-    void segmentPinyinNoSpaces();
-    void segmentPinyinNoDigitsNoSpaces();
-    void segmentPinyinNoDigitsApostrophe();
-    void segmentPinyinDigitsApostrophe();
-    void segmentPinyinRemoveSpecialCharacters();
-    void segmentPinyinKeepGlobCharacters();
-    void segmentPinyinKeepGlobCharactersNoWhitespace();
-    void segmentPinyinKeepMultipleGlobCharacters();
-    void segmentPinyinKeepMultipleGlobCharactersWhitespace();
-    void segmentPinyinKeepMultipleGlobCharactersWhitespaceSurround();
-    void segmentPinyinGlobCharactersTrimWhitespace();
-    void segmentPinyinKeepSpecialCharacters();
-    void segmentPinyinRemoveWhitespace();
-    void segmentPinyinLower();
-    void segmentPinyinLowerWithDigits();
-    void segmentPinyinMultipleFinalsVowelsOnly();
-    void segmentPinyinMultipleFinals();
-    void segmentPinyinInvalidTone();
-    void segmentPinyinGarbage();
 };
 
 TestChineseUtils::TestChineseUtils() {}
@@ -211,520 +101,6 @@ void TestChineseUtils::compareStringsCompatibilityVariantNormalization()
     QCOMPARE(result, Utils::SAME_CHARACTER_STRING);
 }
 
-void TestChineseUtils::jyutpingToYaleSimple()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("si1 zi2 saan1");
-    qDebug() << result.c_str();
-    QCOMPARE(result, "sī jí sāan");
-}
-
-void TestChineseUtils::jyutpingToYaleRejectNoTone()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("joeng");
-    QCOMPARE(result, "joeng");
-}
-
-void TestChineseUtils::jyutpingToYaleRejectSingleLetter()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("a");
-    QCOMPARE(result, "a");
-}
-
-void TestChineseUtils::jyutpingToYaleRejectSpecialCharacter()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("-");
-    QCOMPARE(result, "x");
-}
-
-void TestChineseUtils::jyutpingToYaleNoSpaces()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("si1zi2saan1");
-    QCOMPARE(result, "sī jí sāan");
-}
-
-void TestChineseUtils::jyutpingToYaleSpacesToSegment()
-{
-    std::string result
-        = CantoneseUtils::convertJyutpingToYale("si1 zi2 saan1",
-                                              /* usSpacesToSegment = */ true);
-    QCOMPARE(result, "sī jí sāan");
-}
-
-void TestChineseUtils::jyutpingToYaleSpecialFinal()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("goek3jyun5");
-    QCOMPARE(result, "geuk yúhn");
-}
-
-void TestChineseUtils::jyutpingToYaleLightTone()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("lok6 jyu5");
-    QCOMPARE(result, "lohk yúh");
-}
-
-void TestChineseUtils::jyutpingToYaleSpecialSyllable()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("m4 hai6");
-    QCOMPARE(result, "m̀h haih");
-}
-
-void TestChineseUtils::jyutpingToYaleTones()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale(
-        "saam1 gau2 sei3 ling4 ng5 ji6 cat1 baat3 luk6");
-    QCOMPARE(result, "sāam gáu sei lìhng ńgh yih chāt baat luhk");
-}
-
-void TestChineseUtils::jyutpingToYaleNoTone()
-{
-    std::string result = CantoneseUtils::convertJyutpingToYale("mit");
-    QCOMPARE(result, "mit");
-}
-
-void TestChineseUtils::jyutpingToIPASimple()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("joeng4 sing4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "jœ̽ːŋ ˨ ˩  sɪŋ ˨ ˩");
-#else
-    QCOMPARE(result, "jœ̽ːŋ˨˩  sɪŋ˨˩");
-#endif
-}
-
-void TestChineseUtils::jyutpingToIPARejectNoTone()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("joeng");
-    QCOMPARE(result, "joeng");
-}
-
-void TestChineseUtils::jyutpingToIPARejectSingleLetter()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("a");
-    QCOMPARE(result, "a");
-}
-
-void TestChineseUtils::jyutpingToIPARejectSpecialCharacter()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("-");
-    QCOMPARE(result, "x");
-}
-
-void TestChineseUtils::jyutpingToIPANoSpaces()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("faa1sing4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "fäː ˥  sɪŋ ˨ ˩");
-#else
-    QCOMPARE(result, "fäː˥  sɪŋ˨˩");
-#endif
-}
-
-void TestChineseUtils::jyutpingToIPASpacesToSegment()
-{
-    std::string result
-        = CantoneseUtils::convertJyutpingToIPA("joeng4 sing4",
-                                             /* useSpacesToSegment = */ true);
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "jœ̽ːŋ ˨ ˩  sɪŋ ˨ ˩");
-#else
-    QCOMPARE(result, "jœ̽ːŋ˨˩  sɪŋ˨˩");
-#endif
-}
-
-void TestChineseUtils::jyutpingToIPAPreprocessInitial()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("zyu2 sung3");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "t͡ʃyː ˧ ˥  sʊŋ ˧");
-#else
-    QCOMPARE(result, "t͡ʃyː˧˥  sʊŋ˧");
-#endif
-}
-
-void TestChineseUtils::jyutpingToIPASpecialSyllable()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("m4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "m̩ ˨ ˩");
-#else
-    QCOMPARE(result, "m̩˨˩");
-#endif
-}
-
-void TestChineseUtils::jyutpingToIPACheckedTone()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA(
-        "sik6 si2 o1 faan6");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "sɪk̚ ˨  siː ˧ ˥  ɔː ˥  fäːn ˨");
-#else
-    QCOMPARE(result, "sɪk̚˨  siː˧˥  ɔː˥  fäːn˨");
-#endif
-}
-
-void TestChineseUtils::jyutpingToIPASpecialFinal()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("uk1 kei2 jan4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "ʊk̚ ˥  kʰei̯ ˧ ˥  jɐn ˨ ˩");
-#else
-    QCOMPARE(result, "ʊk̚˥  kʰei̯˧˥  jɐn˨˩");
-#endif
-}
-void TestChineseUtils::jyutpingToIPATones()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA(
-        "saam1 gau2 sei3 ling4 ng5 ji6 cat1 baat3 luk6");
-#ifdef Q_OS_MAC
-    QCOMPARE(result,
-             "säːm ˥  kɐu̯ ˧ ˥  sei̯ ˧  lɪŋ ˨ ˩  ŋ̍ ˩ ˧  jiː ˨  t͡sʰɐt̚ ˥  päːt̚ ˧  "
-             "lʊk̚ ˨");
-#else
-    QCOMPARE(result,
-             "säːm˥  kɐu̯˧˥  sei̯˧  lɪŋ˨˩  ŋ̍˩˧  jiː˨  t͡sʰɐt̚˥  päːt̚˧  lʊk̚˨");
-#endif
-}
-void TestChineseUtils::jyutpingToIPANoTone()
-{
-    std::string result = CantoneseUtils::convertJyutpingToIPA("mok");
-    QCOMPARE(result, "mok");
-}
-
-void TestChineseUtils::prettyPinyinSimple()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("shuai4 ge1");
-    QCOMPARE(result, "shuài gē");
-}
-
-void TestChineseUtils::prettyPinyinRejectNoTone()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("ba");
-    QCOMPARE(result, "ba");
-}
-
-void TestChineseUtils::prettyPinyinRejectSingleLetter()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("a");
-    QCOMPARE(result, "a");
-}
-
-void TestChineseUtils::prettyPinyinRejectSpecialCharacter()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("-");
-    QCOMPARE(result, "-");
-}
-
-void TestChineseUtils::prettyPinyinSecondaryVowel()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("hui4 tu2");
-    QCOMPARE(result, "huì tú");
-}
-
-void TestChineseUtils::prettyPinyinUmlaut()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("nu:3 hai2");
-    QCOMPARE(result, "nǚ hái");
-}
-
-void TestChineseUtils::prettyPinyinTones()
-{
-    std::string result = MandarinUtils::createPrettyPinyin(
-        "ma1 ma2 ma3 ma4 ma5");
-    QCOMPARE(result, "mā má mǎ mà ma");
-}
-void TestChineseUtils::prettyPinyinNoTone()
-{
-    std::string result = MandarinUtils::createPrettyPinyin("nu");
-    QCOMPARE(result, "nu");
-}
-
-void TestChineseUtils::numberedPinyinSimple()
-{
-    std::string result = MandarinUtils::createNumberedPinyin("nu:3 hai2");
-    QCOMPARE(result, "nü3 hai2");
-}
-
-void TestChineseUtils::pinyinWithVSimple()
-{
-    std::string result = MandarinUtils::createPinyinWithV("nu:3 hai2");
-    QCOMPARE(result, "nv3 hai2");
-}
-
-void TestChineseUtils::pinuyinToZhuyinSimple()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("ba1 da2 tong1");
-    QCOMPARE(result, "ㄅㄚ ㄉㄚˊ ㄊㄨㄥ");
-}
-
-void TestChineseUtils::pinuyinToZhuyinRejectNoTone()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("ba");
-    QCOMPARE(result, "ba");
-}
-
-void TestChineseUtils::pinuyinToZhuyinRejectSingleLetter()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("a");
-    QCOMPARE(result, "a");
-}
-void TestChineseUtils::pinuyinToZhuyinRejectSpecialCharacter()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("-");
-    QCOMPARE(result, "-");
-}
-void TestChineseUtils::pinuyinToZhuyinNoSpaces()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("ba1da2tong1");
-    QCOMPARE(result, "ㄅㄚ ㄉㄚˊ ㄊㄨㄥ");
-}
-void TestChineseUtils::pinuyinToZhuyinSpacesToSegment()
-{
-    std::string result
-        = MandarinUtils::convertPinyinToZhuyin("ba1 da2 tong1",
-                                               /* useSpacesToSegment = */ true);
-    QCOMPARE(result, "ㄅㄚ ㄉㄚˊ ㄊㄨㄥ");
-}
-void TestChineseUtils::pinuyinToZhuyinSpecialInitial()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("qu4");
-    QCOMPARE(result, "ㄑㄩˋ");
-
-    result = MandarinUtils::convertPinyinToZhuyin("chi1");
-    QCOMPARE(result, "ㄔ");
-
-    result = MandarinUtils::convertPinyinToZhuyin("ri4");
-    QCOMPARE(result, "ㄖˋ");
-}
-void TestChineseUtils::pinuyinToZhuyinSpecialFinals()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("hm5");
-    QCOMPARE(result, "˙ㄏㄇ");
-
-    result = MandarinUtils::convertPinyinToZhuyin("hng5");
-    QCOMPARE(result, "˙ㄏㄫ");
-
-    result = MandarinUtils::convertPinyinToZhuyin("er2");
-    QCOMPARE(result, "ㄦˊ");
-}
-void TestChineseUtils::pinuyinToZhuyinErhua()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("quanr1");
-    QCOMPARE(result, "ㄑㄩㄢㄦ");
-}
-void TestChineseUtils::pinuyinToZhuyinMalformed()
-{
-    std::string result = MandarinUtils::convertPinyinToZhuyin("chzng2 quanr1");
-    QCOMPARE(result, "chzng2 ㄑㄩㄢㄦ");
-}
-
-void TestChineseUtils::pinyinToIPASimple()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("ba1 da2 tong1");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "pä ˥ ˥  tä ˧ ˥  tʰʊŋ ˥ ˥");
-#else
-    QCOMPARE(result, "pä˥˥  tä˧˥  tʰʊŋ˥˥");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPARejectNoTone()
-{
-    std::string result = MandarinUtils::convertPinyinToIPA("ba");
-    QCOMPARE(result, "ba");
-}
-
-void TestChineseUtils::pinyinToIPARejectSingleLetter()
-{
-    std::string result = MandarinUtils::convertPinyinToIPA("a");
-    QCOMPARE(result, "a");
-}
-
-void TestChineseUtils::pinyinToIPARejectSpecialCharacter()
-{
-    std::string result = MandarinUtils::convertPinyinToIPA("-");
-    QCOMPARE(result, "-");
-}
-
-void TestChineseUtils::pinyinToIPANoSpaces()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("ba1da2tong1");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "pä ˥ ˥  tä ˧ ˥  tʰʊŋ ˥ ˥");
-#else
-    QCOMPARE(result, "pä˥˥  tä˧˥  tʰʊŋ˥˥");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPASpacesToSegment()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result
-        = MandarinUtils::convertPinyinToIPA("ba1 da2 tong1",
-                                            /* useSpacesToSegment = */ true);
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "pä ˥ ˥  tä ˧ ˥  tʰʊŋ ˥ ˥");
-#else
-    QCOMPARE(result, "pä˥˥  tä˧˥  tʰʊŋ˥˥");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPASpecialCaseNg()
-{
-    std::string result = MandarinUtils::convertPinyinToIPA("ng5");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "ŋ̍ ");
-#else
-    QCOMPARE(result, "ŋ̍");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPASpecialCaseRi()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("ri4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "ʐ̩ ˥ ˩");
-#else
-    QCOMPARE(result, "ʐ̩˥˩");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPASyllableWithV()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("nv3");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "ny ˨ ˩ ˦");
-#else
-    QCOMPARE(result, "ny˨˩˦");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("qu4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "t͡ɕʰy ˥ ˩");
-#else
-    QCOMPARE(result, "t͡ɕʰy˥˩");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPAVoicelessInitial()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("ge5");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "g̊ə ");
-#else
-    QCOMPARE(result, "g̊ə");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("yi1 ge5");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "i ˥ ˥  g̊ə ˨");
-#else
-    QCOMPARE(result, "i˥˥  g̊ə˨");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPAToneThree()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("ke3");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "kʰɤ ˨ ˩ ˦");
-#else
-    QCOMPARE(result, "kʰɤ˨˩˦");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("ke3 yi3");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "kʰɤ ˨ ˩ ˦ ꜔ ꜒  i ˨ ˩ ˦ ꜕ ꜖ ( ꜓ )");
-#else
-    QCOMPARE(result, "kʰɤ˨˩˦꜔꜒  i˨˩˦꜕꜖(꜓)");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPAToneFour()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("xia4 qu4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "ɕjä ˥ ˩ ꜒ ꜔  t͡ɕʰy ˥ ˩");
-#else
-    QCOMPARE(result, "ɕjä˥˩꜒꜔  t͡ɕʰy˥˩");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("xia4");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "ɕjä ˥ ˩");
-#else
-    QCOMPARE(result, "ɕjä˥˩");
-#endif
-}
-
-void TestChineseUtils::pinyinToIPAOtherTone()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("ma1");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "mä ˥ ˥");
-#else
-    QCOMPARE(result, "mä˥˥");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("ma2");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "mä ˧ ˥");
-#else
-    QCOMPARE(result, "mä˧˥");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("ma5");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "mä ");
-#else
-    QCOMPARE(result, "mä");
-#endif
-}
-void TestChineseUtils::pinyinToIPAErhua()
-{
-#ifdef Q_OS_WINDOWS
-    QSKIP("IPA tone comparison doesn't work on Windows, skipping test");
-#endif
-    std::string result = MandarinUtils::convertPinyinToIPA("huar1");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "xu̯ɑɻ ˥ ˥");
-#else
-    QCOMPARE(result, "xu̯ɑɻ˥˥");
-#endif
-
-    result = MandarinUtils::convertPinyinToIPA("quanr1");
-#ifdef Q_OS_MAC
-    QCOMPARE(result, "t͡ɕʰɥɑɻ ˥ ˥");
-#else
-    QCOMPARE(result, "t͡ɕʰɥɑɻ˥˥");
-#endif
-}
-
 void TestChineseUtils::constructRomanisationQuerySingleSyllable()
 {
     std::string result = ChineseUtils::constructRomanisationQuery({"se"}, "?");
@@ -780,489 +156,6 @@ void TestChineseUtils::constructRomanisationQueryOnlyGlobCharacters()
     std::string result = ChineseUtils::constructRomanisationQuery(
         {"?", "?", "?", "?", "? ", "?", "?", "?", "?"}, "?");
     QCOMPARE(QString::fromStdString(result), "????? ????");
-}
-
-void TestChineseUtils::segmentJyutpingSimple()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m4 goi1", result);
-    std::vector<std::string> expected = {"m4", "goi1"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingNoDigits()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m goi", result);
-    std::vector<std::string> expected = {"m", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingNoSpaces()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m4goi1", result);
-    std::vector<std::string> expected = {"m4", "goi1"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingNoDigitsNoSpaces()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("mgoi", result);
-    std::vector<std::string> expected = {"m", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingNoDigitsApostrophe()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m'aam", result);
-    std::vector<std::string> expected = {"m", "aam"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingDigitsApostrophe()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m4'aam", result);
-    std::vector<std::string> expected = {"m4", "aam"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingRemoveSpecialCharacters()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m*goi", result);
-    std::vector<std::string> expected = {"m", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingKeepGlobCharacters()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m* goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", "* ", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingKeepGlobCharactersNoWhitespace()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m*goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", "*", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingKeepMultipleGlobCharacters()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m?* goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", "?", "* ", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingKeepMultipleGlobCharactersWhitespace()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m? * goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", "? ", "* ", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingKeepMultipleGlobCharactersWhitespaceSurround()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m ? * goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", " ? ", "* ", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingGlobCharactersTrimWhitespace()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m  ?            *      goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", " ? ", "* ", "goi"};
-    QCOMPARE(result, expected);
-
-    CantoneseUtils::segmentJyutping("m?* ????",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    expected = {"m", "?", "* ", "?", "?", "?", "?"};
-    QCOMPARE(result, expected);
-
-    CantoneseUtils::segmentJyutping("m * ????*",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    expected = {"m", " * ", "?", "?", "?", "?", "*"};
-    QCOMPARE(result, expected);
-
-    CantoneseUtils::segmentJyutping("m? goi*",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    expected = {"m", "? ", "goi", "*"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingKeepSpecialCharacters()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m？ goi",
-                                  result,
-                                  /* removeSpecialCharacters = */ false);
-    std::vector<std::string> expected = {"m", "？", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingRemoveWhitespace()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("  m                           goi      ",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"m", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingLower()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("mGoI", result);
-    std::vector<std::string> expected = {"m", "goi"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingLowerWithDigits()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("m4GoI1", result);
-    std::vector<std::string> expected = {"m4", "goi1"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingMultipleFinalsVowelsOnly()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("aaaa",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"aa", "aa"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingMultipleFinalsConsonantsOnly()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("ngng",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"ng", "ng"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingMultipleFinals()
-{
-    std::vector<std::string> result;
-    CantoneseUtils::segmentJyutping("amam",
-                                  result,
-                                  /* removeSpecialCharacters = */ true,
-                                  /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"am", "am"};
-    QCOMPARE(result, expected);
-}
-
-void TestChineseUtils::segmentJyutpingInvalidTone()
-{
-    std::vector<std::string> result;
-    bool valid
-        = MandarinUtils::segmentPinyin("hang0",
-                                       result,
-                                       /* removeSpecialCharacters = */ true,
-                                       /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"hang0"};
-    QCOMPARE(result, expected);
-    QCOMPARE(valid, false);
-
-    valid = MandarinUtils::segmentPinyin("hang7",
-                                         result,
-                                         /* removeSpecialCharacters = */ true,
-                                         /* removeGlobCharacters = */ false);
-    expected = {"hang7"};
-    QCOMPARE(result, expected);
-    QCOMPARE(valid, false);
-}
-
-void TestChineseUtils::segmentJyutpingGarbage()
-{
-    std::vector<std::string> result;
-    bool valid
-        = CantoneseUtils::segmentJyutping("kljnxclkjvnl",
-                                        result,
-                                        /* removeSpecialCharacters = */ true,
-                                        /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"kljnxclkjvnl"};
-    QCOMPARE(result, expected);
-    QCOMPARE(valid, false);
-}
-
-void TestMandarinUtils::segmentPinyinSimple()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang3 dong1", result);
-    std::vector<std::string> expected = {"guang3", "dong1"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinNoDigits()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang dong", result);
-    std::vector<std::string> expected = {"guang", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinNoSpaces()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang3dong1", result);
-    std::vector<std::string> expected = {"guang3", "dong1"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinNoDigitsNoSpaces()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guangdong", result);
-    std::vector<std::string> expected = {"guang", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinNoDigitsApostrophe()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("xi'an", result);
-    std::vector<std::string> expected = {"xi", "an"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinDigitsApostrophe()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("xi1'an", result);
-    std::vector<std::string> expected = {"xi1", "an"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinRemoveSpecialCharacters()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang*dong!!", result);
-    std::vector<std::string> expected = {"guang", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinKeepGlobCharacters()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang* dong?",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", "* ", "dong", "?"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinKeepGlobCharactersNoWhitespace()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang*dong?",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", "*", "dong", "?"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinKeepMultipleGlobCharacters()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang?* dong",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", "?", "* ", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinKeepMultipleGlobCharactersWhitespace()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang? * dong",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", "? ", "* ", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinKeepMultipleGlobCharactersWhitespaceSurround()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang ? * dong",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", " ? ", "* ", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinGlobCharactersTrimWhitespace()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang  ?            *      dong",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", " ? ", "* ", "dong"};
-    QCOMPARE(result, expected);
-
-    MandarinUtils::segmentPinyin("guang?* ?????",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    expected = {"guang", "?", "* ", "?", "?", "?", "?", "?"};
-    QCOMPARE(result, expected);
-
-    MandarinUtils::segmentPinyin("guang * ????*",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    expected = {"guang", " * ", "?", "?", "?", "?", "*"};
-    QCOMPARE(result, expected);
-
-    MandarinUtils::segmentPinyin("guang? dong*",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    expected = {"guang", "? ", "dong", "*"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinKeepSpecialCharacters()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guang？ dong1",
-                                 result,
-                                 /* removeSpecialCharacters = */ false);
-    std::vector<std::string> expected = {"guang", "？", "dong1"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinRemoveWhitespace()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("  guang                           dong      ",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"guang", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinLower()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("gUanGdOnG", result);
-    std::vector<std::string> expected = {"guang", "dong"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinLowerWithDigits()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("guAng3dONg1", result);
-    std::vector<std::string> expected = {"guang3", "dong1"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinMultipleFinalsVowelsOnly()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("ee",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"e", "e"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinMultipleFinals()
-{
-    std::vector<std::string> result;
-    MandarinUtils::segmentPinyin("angang",
-                                 result,
-                                 /* removeSpecialCharacters = */ true,
-                                 /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"ang", "ang"};
-    QCOMPARE(result, expected);
-}
-
-void TestMandarinUtils::segmentPinyinInvalidTone()
-{
-    std::vector<std::string> result;
-    bool valid
-        = MandarinUtils::segmentPinyin("heng0",
-                                       result,
-                                       /* removeSpecialCharacters = */ true,
-                                       /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"heng0"};
-    QCOMPARE(result, expected);
-    QCOMPARE(valid, false);
-
-    valid = MandarinUtils::segmentPinyin("heng6",
-                                         result,
-                                         /* removeSpecialCharacters = */ true,
-                                         /* removeGlobCharacters = */ false);
-    expected = {"heng6"};
-    QCOMPARE(result, expected);
-    QCOMPARE(valid, false);
-}
-
-void TestMandarinUtils::segmentPinyinGarbage()
-{
-    std::vector<std::string> result;
-    bool valid
-        = MandarinUtils::segmentPinyin("kljnxclkjvnl",
-                                       result,
-                                       /* removeSpecialCharacters = */ true,
-                                       /* removeGlobCharacters = */ false);
-    std::vector<std::string> expected = {"kljnxclkjvnl"};
-    QCOMPARE(result, expected);
-    QCOMPARE(valid, false);
 }
 
 QTEST_APPLESS_MAIN(TestChineseUtils)
